Create README.md
This is a Python-based project. Using Python, Django, and SQLite, It is a payroll management system.
The key module in this project is the employee module, which handles the employee features.
It keeps track of all of the employee’s information and data. We’ve created all of the employees, and company crud (create, read, update, and delete) operations